Doing "ReactApp - Tic Tac Toe" - Finally, the login with the API is working
