--- react/react-training/src/app/screens/Login/components/SignInForm/index.js
@@ -1,24 +1,0 @@
-import React, { Component } from 'react';
-import { reduxForm } from 'redux-form';
-import PropTypes from 'prop-types';
-
-import { FORM_SIGIN } from '@constants';
-
-import Form from './layout';
-
-class SignInForm extends Component {
-  render() {
-    const { handleSubmit, userLoginError, userIsLogged } = this.props;
-    return <Form handleSubmit={handleSubmit} userIsLogged={userIsLogged} userLoginError={userLoginError} />;
-  }
-}
-
-SignInForm.propTypes = {
-  handleSubmit: PropTypes.func.isRequired,
-  userIsLogged: PropTypes.bool,
-  userLoginError: PropTypes.string
-};
-
-export default reduxForm({
-  form: FORM_SIGIN
-})(SignInForm);
